Move AddressBook, ReadOnlyAddressBook ReadOnlyUserPrefs UserPrefs outside of person package
