+++ src/main/java/seedu/address/logic/LogicPersonManager.java
@@ -13,7 +13,7 @@ import seedu.address.logic.commands.person.Command;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.logic.parser.person.AddressBookParser;
 import seedu.address.model.ModelPerson;
-import seedu.address.model.person.ReadOnlyAddressBook;
+import seedu.address.model.ReadOnlyAddressBook;
 import seedu.address.model.person.person.Person;
 import seedu.address.storage.StoragePerson;
