+++ src/main/java/seedu/address/model/ModelPerson.java
@@ -5,8 +5,6 @@ import java.util.function.Predicate;
 
 import javafx.collections.ObservableList;
 import seedu.address.commons.core.GuiSettings;
-import seedu.address.model.person.ReadOnlyAddressBook;
-import seedu.address.model.person.ReadOnlyUserPrefs;
 import seedu.address.model.person.person.Person;
 
 /**
