+++ src/test/java/seedu/address/logic/LogicPersonManagerTest.java
@@ -24,8 +24,8 @@ import seedu.address.logic.commands.person.ListCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.ModelPerson;
 import seedu.address.model.ModelPersonManager;
-import seedu.address.model.person.ReadOnlyAddressBook;
-import seedu.address.model.person.UserPrefs;
+import seedu.address.model.ReadOnlyAddressBook;
+import seedu.address.model.UserPrefs;
 import seedu.address.model.person.person.Person;
 import seedu.address.storage.StoragePersonManager;
 import seedu.address.storage.person.JsonAddressBookStorage;
