+++ src/test/java/seedu/address/logic/commands/person/AddCommandIntegrationTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.model.ModelPerson;
 import seedu.address.model.ModelPersonManager;
-import seedu.address.model.person.UserPrefs;
+import seedu.address.model.UserPrefs;
 import seedu.address.model.person.person.Person;
 import seedu.address.testutil.PersonBuilder;
