+++ src/test/java/seedu/address/logic/commands/person/ClearCommandTest.java
@@ -7,8 +7,8 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.model.ModelPerson;
 import seedu.address.model.ModelPersonManager;
-import seedu.address.model.person.AddressBook;
-import seedu.address.model.person.UserPrefs;
+import seedu.address.model.AddressBook;
+import seedu.address.model.UserPrefs;
 
 public class ClearCommandTest {
