+++ src/test/java/seedu/address/logic/commands/person/DeleteCommandTest.java
@@ -15,7 +15,7 @@ import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
 import seedu.address.model.ModelPerson;
 import seedu.address.model.ModelPersonManager;
-import seedu.address.model.person.UserPrefs;
+import seedu.address.model.UserPrefs;
 import seedu.address.model.person.person.Person;
 
 /**
