+++ src/test/java/seedu/address/logic/commands/person/EditCommandTest.java
@@ -21,8 +21,8 @@ import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.person.EditCommand.EditPersonDescriptor;
 import seedu.address.model.ModelPerson;
 import seedu.address.model.ModelPersonManager;
-import seedu.address.model.person.AddressBook;
-import seedu.address.model.person.UserPrefs;
+import seedu.address.model.AddressBook;
+import seedu.address.model.UserPrefs;
 import seedu.address.model.person.person.Person;
 import seedu.address.testutil.EditPersonDescriptorBuilder;
 import seedu.address.testutil.PersonBuilder;
