+++ src/test/java/seedu/address/logic/commands/person/FindCommandTest.java
@@ -17,7 +17,7 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.model.ModelPerson;
 import seedu.address.model.ModelPersonManager;
-import seedu.address.model.person.UserPrefs;
+import seedu.address.model.UserPrefs;
 import seedu.address.model.person.person.NameContainsKeywordsPredicate;
 
 /**
