+++ src/test/java/seedu/address/model/AddressBookTest.java
@@ -18,8 +18,6 @@ import org.junit.jupiter.api.Test;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
-import seedu.address.model.person.AddressBook;
-import seedu.address.model.person.ReadOnlyAddressBook;
 import seedu.address.model.person.person.Person;
 import seedu.address.model.person.person.exceptions.DuplicatePersonException;
 import seedu.address.testutil.PersonBuilder;
