+++ src/test/java/seedu/address/model/ModelPersonManagerTest.java
@@ -15,8 +15,6 @@ import java.util.Arrays;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.GuiSettings;
-import seedu.address.model.person.AddressBook;
-import seedu.address.model.person.UserPrefs;
 import seedu.address.model.person.person.NameContainsKeywordsPredicate;
 import seedu.address.testutil.AddressBookBuilder;
