+++ src/test/java/seedu/address/storage/person/JsonUserPrefsStoragePersonTest.java
@@ -14,7 +14,7 @@ import org.junit.jupiter.api.io.TempDir;
 
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.commons.exceptions.DataConversionException;
-import seedu.address.model.person.UserPrefs;
+import seedu.address.model.UserPrefs;
 
 public class JsonUserPrefsStoragePersonTest {
