+++ src/test/java/seedu/address/testutil/TypicalPersons.java
@@ -15,7 +15,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-import seedu.address.model.person.AddressBook;
+import seedu.address.model.AddressBook;
 import seedu.address.model.person.person.Person;
 
 /**
